Change layout, add flow meter and make monitoring more robust

+++ lib/objects/Objects.h
@@ -19,7 +19,7 @@ typedef struct {
 
 typedef struct {
     double rawVoltage;
-    double rawFlow;
+    byte rawFlow;
     byte percentageVoltage;
     byte percentageFlow;
 } monitor_t;
